fix: restore project settings ownership
